+++ recipes/cron.rb
@@ -1,20 +1,26 @@
-
+#
 # Cookbook Name:: ec2-consistent-snapshot
 # Recipe:: cron
-include_recipe 'cron::default'
+#
+include_recipe 'cron'
 
-cron = node['ec2-consistent-snapshot']['cron']
+script = ec2_consistent_snapshot_script node['ec2-consistent-snapshot']['script']['name'] do
+  mysql node['ec2-consistent-snapshot']['script']['mysql']
+  mysql_socket node['ec2-consistent-snapshot']['script']['mysql_socket']
+  freeze_filesystem node['ec2-consistent-snapshot']['script']['freeze_filesystem']
+  aws_credentials_file node['ec2-consistent-snapshot']['script']['aws_credentials_file']
+  mysql_master_status_file node['ec2-consistent-snapshot']['script']['mysql_master_status_file']
+  mysql_host node['ec2-consistent-snapshot']['script']['mysql_host']
+  mysql_username node['ec2-consistent-snapshot']['script']['mysql_username']
+  mysql_password node['ec2-consistent-snapshot']['script']['mysql_password']
+  description node['ec2-consistent-snapshot']['script']['description']
+  volumes node['ec2-consistent-snapshot']['script']['volumes']
+end
 
-cron_d "ec2-consistent-snapshot" do
-  minute cron['minute']
-  hour cron['hour']
-  day cron['day']
-  month cron['month']
-  weekday cron['weekday']
-  command cron['command']
-  user cron['user']
-  mailto cron['mailto']
-  path cron['path']
-  home cron['home']
-  shell cron['shell']
+cron_d node['ec2-consistent-snapshot']['cron']['name'] do
+    minute node['ec2-consistent-snapshot']['cron']['minute']
+    hour node['ec2-consistent-snapshot']['cron']['hour']
+    command script.name
+    user node['ec2-consistent-snapshot']['user']
+    mailto node['ec2-consistent-snapshot']['mailto']
 end